requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
aiosqlite==0.20.0
pyjwt==2.9.0
bcrypt==4.2.0
python-multipart==0.0.9
httpx==0.27.0
apscheduler==3.10.4
